게시글 댓글 개수 관리 로직

+++ mvc-api/src/main/kotlin/com/jinuk/toy/mvcapi/view/post/response/PostSearchResponse.kt
@@ -9,7 +9,7 @@ class PostSearchResponse(
     val id: Long,
     val title: PostTitle,
     val userName: Username,
-    val commentCount: Int,
+    val commentCount: Long,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime,
 )
